==> studies/study1_sparse/4_study1_sparse_results.qmd <==
---
title: "Analyzing study 1 sparse results"
author: "Jenna Landy"
---

This sensitivty analysis shows that even with sparse data (matrices are 20-40% sparse), the Poisson-Exponential + MH model is still able to match the standard Poisson-Exponential model, both in terms of MAP estimates and posterior uncertainty. This analysis is done only looking at N = 4 signatures and only with Exponential priors.

This analysis assumes `study1_sparse_processing.R` has been run, and that the metrics are in `processed/study1_sparse/metrics.csv`.

```{r}
library(bayesNMF)
library(tidyverse)
library(ggplot2)
library(ggridges)
library(grid)
library(gridExtra)
library(glue)
library(cowplot)
source("../study1_visualization_functions.R")

data_dir <- "../../data/study1_sparse"
output_dir <- "../../output/study1_sparse"
processed_dir <- "../../processed/study1_sparse"
figures_dir <- "../../figures/study1_sparse"
```

The `metrics` dataframe holds metrics comparing each result to ground truth.
```{r}
metrics <- read_csv(file.path(processed_dir, "metrics.csv"))
metrics %>%
  group_by(N, G, likelihood, prior, MH) %>%
  tally() %>%
  arrange(n)
```

The following dataframes hold metrics comparing each result to the other results for the same dataset. Models are compared to others with the same prior (Gamma is compared to the Truncated Normal).
```{r}
Standard_vs_MH_MAPs <- read_csv(file.path(processed_dir, "Standard_vs_MH_MAPs.csv"))
head(Standard_vs_MH_MAPs)

Standard_vs_Standard_repeated_MAPs <- read_csv(file.path(processed_dir, "Standard_vs_Standard_repeated_MAPs.csv"))
head(Standard_vs_Standard_repeated_MAPs)

Standard_vs_MH_CIs <- read_csv(file.path(processed_dir, "Standard_vs_MH_CIs.csv"))
head(Standard_vs_MH_CIs)
colnames(Standard_vs_MH_CIs)
```

## All Metrics

```{r}
metrics %>%
  pivot_longer(
    c(min_sim, total_minutes, RMSE, KL, iters), 
    names_to = "metric", 
    values_to = "value"
  ) %>%
  arrange(N) %>%
  mutate(
    N = paste(N, "Signatures"),
    N = factor(N, levels = unique(N)),
    metric = case_when(
      metric == "min_sim" ~ "Minimum Cosine Similarity",
      metric == "total_minutes" ~ "Total Minutes",
      metric == "RMSE" ~ "RMSE",
      metric == "KL" ~ "KL Divergence",
      metric == "iters" ~ "Iterations"
    ),
    model = paste(likelihood, prior, sep = "_"),
    model = ifelse(MH, paste(model, "MH", sep = "_"), model),
    model = factor(model, levels = names(color_list))
  ) %>%
  ggplot(aes(x = as.factor(G), y = value, color = model)) +
  facet_grid(rows = vars(metric), cols = vars(N), scales = "free") +
  geom_boxplot() +
  theme_bw() +
  labs(
    x = "Sample Size"
  ) +
  theme(
    axis.title.y = element_blank(),
    legend.position = 'top',
    legend.justification = 'left'
  )
ggsave(file.path(figures_dir, "study1_sparse_metrics.png"), width = 10, height = 10)
```

## Acceptance rates

```{r}
acceptance_rates <- metrics %>%
  filter(MH) %>%
  pivot_longer(cols = c(mean_mean_acceptance_P, mean_mean_acceptance_E), names_to = "which", values_to = "value") %>%
  mutate(
    which = case_when(
      which == "mean_mean_acceptance_P" ~ "Signatures",
      which == "mean_mean_acceptance_E" ~ "Exposures"
    ),
    which = factor(which, levels = c("Signatures", "Exposures"))
  ) %>%
  ggplot(aes(x = factor(G, levels = sort(unique(G))), y = value, color = which)) +
  geom_boxplot() +
  labs(
    x = "Sample Size",
    y = "Mean acceptance rate",
    color = "Matrix"
  ) +
  theme_bw() +
  theme(
    legend.position = "top",
    legend.justification = "left"
  )
ggsave(file.path(figures_dir, "acceptance_rates.png"), width = 10, height = 5)
```

## Comparing MAPs

Result 1: For fixed rank, Normal and Poisson models result in very similar MAP point estimates
Looking at exponential priors first, directly compare the metrics of each dataset between Poisson (non-MH) and Normal models. 
```{r}
threshold <- 0.9
fontsize = 25

comparing_MAPs <- Standard_vs_MH_MAPs %>%
  rbind(Standard_vs_Standard_repeated_MAPs) %>%
  mutate(
    compare = case_when(
      compare == 'Poisson<vs>Poisson_MH' ~ "Poisson vs Poisson + MH",
      compare == 'Poisson<vs>Poisson_repeated' ~ "Poisson vs Poisson (repeated)"
    ),
    compare = factor(compare, levels = c("Poisson vs Poisson + MH", "Poisson vs Poisson (repeated)"))
  ) %>%
  select(N, G, rep, prior, compare, P_min_cosine, E_min_cosine) %>%
  pivot_longer(cols = c(P_min_cosine, E_min_cosine)) %>%
  mutate(
    name = ifelse(
      name == "E_min_cosine",
      "Exposures Matrix (E)",
      "Signatures Matrix (P)"
    )
  ) %>%
  ggplot(aes(x = value, y = prior)) +
  facet_grid(cols = vars(name), rows = vars(compare), switch = 'y') +
  geom_vline(xintercept = 0.99, linetype = "dashed") +
  geom_density_ridges(scale = 4, bandwidth = 0.001) +
  labs(
    x = "Minimum cosine similarity between MAP estimates"
  ) +
  # scale_x_continuous(breaks = seq(0.9, 1, 0.05), limits = c(threshold, 1.01)) +
  theme(
    strip.text.y.left = element_text(angle = 0),
    text = element_text(size = fontsize),
    axis.text.y.left = element_blank(),
    axis.ticks.y.left = element_blank(),
    axis.title.y = element_blank(),
    legend.position = "top",
    legend.justification = "left"
  )
ggsave(file.path(figures_dir, "comparing_MAPs.png"), comparing_MAPs)
```

## Comparing credible intervals

```{r}
dat <- E_10 %>% 
  mutate(name = "10th percentile", type = "Exposures Matrix\nScale = Number of Mutations", diff = E_q10_width_diff) %>%
  select(-E_q10_width_diff) %>%
  rbind(
    E_90 %>% 
    mutate(name = "90th percentile", type = "Exposures Matrix\nScale = Number of Mutations", diff = E_q90_width_diff) %>%
    select(-E_q90_width_diff)
  ) %>%
  rbind(
    P_10 %>% mutate(name = "10th percentile", type = "Signatures Matrix\nScale = Probability", diff = P_q10_width_diff) %>%
    select(-P_q10_width_diff)
  ) %>%
  rbind(
    P_90 %>% 
    mutate(name = "90th percentile", type = "Signatures Matrix\nScale = Probability", diff = P_q90_width_diff) %>%
    select(-P_q90_width_diff)
  ) %>%
  mutate(
    N = paste("N =", N)
  )
  
comparing_widths_combo = ggplot(dat) +
  facet_wrap(vars(type), nrow = 1, scales = 'free_y') +
  geom_boxplot(
    data = dat %>% filter(name == "10th percentile"),fill = "grey", 
    aes(x = as.factor(G), y = diff, alpha = compare, group = g)
  ) +
  geom_boxplot(
    data = dat %>% filter(name == "90th percentile"),fill = "grey", 
    aes(x = as.factor(G), y = diff, alpha = compare, group = g)
  ) +
  labs(
    x = "Sample Size (G)",
    y = "Difference in CI width: Poisson - (Poisson + MH)\n10th percentile (negative) and 90th percentile (positive)",
    alpha = "Comparison"
  ) +
  scale_alpha_manual(
    breaks = c("Poisson<vs>Poisson_repeated", "Poisson<vs>Poisson_MH"),
    labels = c("Poisson<vs>Poisson_repeated" = "Poisson vs Poisson", "Poisson<vs>Poisson_MH" = "Poisson vs Poisson + MH"),
    values = c("Poisson<vs>Poisson_repeated" = 0.1, "Poisson<vs>Poisson_MH" = 1)
  ) +
  guides(
    alpha = guide_legend(override.aes = list(fill = 'black'))
  ) +
  theme_light()

ggsave(file.path(figures_dir, "comparing_CI_widths.png"), comparing_widths_combo, width = 10, height = 5)
```

## Combine into one figure

```{r}
p <- plot_grid(
  comparing_MAPs +
    theme_light(base_size = 12) +
    labs(
      x = "Min cosine similarity between MAP estimates",
    ) +
    ggtitle("A. Agreement of MAP estimates") +
    theme(
      axis.text.y = element_blank(),
      axis.ticks.y = element_blank(),
      axis.title.y = element_blank(),
      strip.text.y.left = element_text(angle = 90),
      strip.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      strip.text = element_text(color = 'black'),
      plot.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      legend.position = "none"
    ), 
  comparing_widths_combo +
    theme_light(base_size = 12) +
    ggtitle("B. Agreement of credible intervals") +
    theme(
      legend.position = "bottom",
      strip.text = element_text(color = 'black'),
      strip.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      plot.background = element_rect(fill = "white", color = "black", linewidth = 0.5)
    ), 
  ncol = 2, rel_widths = c(0.8, 1)
)
ggsave(file.path(figures_dir, "sparse_results.png"), p, width = 10, height = 5)
```

## Combine with data figure

```{r}
data_plot <- readRDS(file.path(figures_dir, "sparse_data.rds"))

row_1 <- plot_grid(
  data_plot,
  acceptance_rates +
    ggtitle("C. Average acceptance rates") +
    theme(
      plot.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      plot.title.position = "plot"
    ),
  nrow = 1, rel_widths = c(2, 1)
)

row_2 <- plot_grid(
  comparing_MAPs +
    theme_light(base_size = 12) +
    labs(
      x = "Min cosine similarity",
    ) +
    ggtitle("D. Agreement of MAP estimates") +
    theme(
      axis.text.y = element_blank(),
      axis.ticks.y = element_blank(),
      axis.title.y = element_blank(),
      strip.text.y.left = element_text(angle = 90, size = 9),
      strip.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      strip.text = element_text(color = 'black'),
      plot.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      legend.position = "none",
      plot.title.position = "plot"
    ), 
  comparing_widths_combo +
    theme_light(base_size = 12) +
    ggtitle("E. Agreement of credible intervals") +
    theme(
      legend.position = "bottom",
      strip.text = element_text(color = 'black'),
      strip.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      plot.background = element_rect(fill = "white", color = "black", linewidth = 0.5),
      plot.title.position = "plot",
      axis.title.y = element_text(size = 10)
    ), 
  ncol = 2, rel_widths = c(0.5, 1)
)

p <- plot_grid(row_1, row_2, ncol = 1, rel_heights = c(0.7, 1))
ggsave(file.path(figures_dir, "study1_sparse.png"), p, width = 10, height = 8)
```
